fix compiler warning

+++ cpp/inline/util/AllocPool.inl
@@ -64,7 +64,7 @@ T& AllocPool<T, N>::operator[](pool_index_t i) {
 
 template <typename T, int N>
 const T& AllocPool<T, N>::operator[](pool_index_t i) const {
-  debug_assert(i >= 0 && i < size_, "Index out of bounds: %ld", i);
+  debug_assert(i >= 0 && size_t(i) < size_, "Index out of bounds: %ld", i);
   int block_index = detail::get_block_index<N>(i);
   int offset = block_index == 0 ? i : (i - std::bit_floor(uint64_t(i)));
